Fix error in the WCS header of re-gridded images
I had originally copied over the original header and then tried to
update it with the new WCS.  But that does not work, since it leaves
behind unwanted keywords from the old WCS, giving an inconsistent
header. I fixed it by just nuking the old header and making an
entirely new one.