Describe helio_utils.py and add necessary dependencies
I previously brought in the file from Alba Orion West project, using a
very simple trick involving email patches:

In Orion West repo:

$ git log --pretty=email --patch-with-stat --reverse --full-index \
      --binary -- helio_utils.py > patch

Then copy patch to this repo and query replace in the patch to put in
the correct path: /src/mes_longslit/helio_utils.py

Then run this in this repo

$ git am --committer-date-is-author-date < patch

To apply the commits. These end up logically at the end of the chain
of commits, but with dates in the past.

Then delete the patch file.